Add files via upload
This notebook uses a cascading selection process to select stocks for a competitive portfolio. It uses either the Nasdaq100 or S&P500, depending on the input. It can choose stocks from a variety of dates that are separated by month, since 2010. The returns of the selected portfolio are then calculated and displayed and compared with the returns of the chosen benchmark since the date that was chosen. This was a simple strategy that tested a cascading selection process and utilized momentum as its main marker for selection.